+++ src/types/index.ts
@@ -27,3 +27,4 @@ export * from './achievement';
 export * from './ranked-battles';
 export * from './clan-capital-league';
 export * from './builder-base-league';
+export * from './forge';